# doc/agent/backend/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "慧农金融AI智能体接口",
    "description": "AI智能体审批工作流相关接口，支持Dify平台调用",
    "version": "1.0.0",
    "contact": {
      "name": "慧农金融技术支持",
      "url": "http://localhost:8080"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    },
    {
      "url": "http://172.18.120.57:8080",
      "description": "内网测试环境"
    }
  ],
  "paths": {
    "/api/v1/ai-agent/applications/{application_id}/info": {
      "get": {
        "summary": "获取申请信息",
        "description": "获取贷款申请的详细信息用于AI分析",
        "operationId": "getApplicationInfo",
        "tags": ["AI智能体"],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "申请ID",
            "example": "app_20241201_001"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取申请信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "成功响应示例",
                    "value": {
                      "code": 0,
                      "message": "获取成功",
                      "data": {
                        "application_id": "app_20241201_001",
                        "product_info": {
                          "product_id": "loan_001",
                          "name": "农户经营贷",
                          "max_amount": 500000,
                          "interest_rate_yearly": "6.5%"
                        },
                        "applicant_info": {
                          "user_id": "user_001",
                          "real_name": "张三",
                          "credit_score": 750,
                          "annual_income": 120000
                        },
                        "amount": 50000,
                        "term_months": 12,
                        "purpose": "农业种植资金周转"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "申请不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AIAgentToken": []
          }
        ]
      }
    },
    "/api/v1/ai-agent/external-data": {
      "get": {
        "summary": "获取外部数据",
        "description": "获取征信、银行流水等外部数据用于AI分析",
        "operationId": "getExternalData",
        "tags": ["AI智能体"],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "用户ID",
            "example": "user_001"
          },
          {
            "name": "data_types",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "数据类型，逗号分隔：credit_report,bank_flow,blacklist_check",
            "example": "credit_report,bank_flow,blacklist_check"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取外部数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "成功响应示例",
                    "value": {
                      "code": 0,
                      "message": "获取成功",
                      "data": {
                        "credit_report": {
                          "score": 750,
                          "level": "优秀",
                          "detail": "信用记录良好，无逾期记录"
                        },
                        "bank_flow": {
                          "average_monthly_income": 15000,
                          "debt_to_income_ratio": 0.25,
                          "account_stability": "稳定"
                        },
                        "blacklist_check": {
                          "is_blacklisted": false,
                          "risk_level": "低风险"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AIAgentToken": []
          }
        ]
      }
    },
    "/api/v1/ai-agent/applications/{application_id}/ai-decision": {
      "post": {
        "summary": "提交AI决策结果",
        "description": "Dify工作流完成分析后，提交AI决策结果",
        "operationId": "submitAIDecision",
        "tags": ["AI智能体"],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "申请ID",
            "example": "app_20241201_001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIDecisionRequest"
              },
              "examples": {
                "auto_approved": {
                  "summary": "自动通过示例",
                  "value": {
                    "decision": "AUTO_APPROVED",
                    "risk_score": 0.25,
                    "risk_level": "LOW",
                    "confidence": 0.85,
                    "analysis_result": {
                      "credit_analysis": "信用分数750，信用记录优秀",
                      "income_analysis": "年收入12万，申请金额5万，收入充足",
                      "risk_factors": [],
                      "approval_amount": 50000,
                      "recommended_rate": "年化利率6.5%"
                    },
                    "processing_info": {
                      "ai_model_version": "v1.2.0",
                      "processing_time_ms": 2500,
                      "workflow_id": "wf_20241201_001"
                    }
                  }
                },
                "require_human_review": {
                  "summary": "需要人工审核示例",
                  "value": {
                    "decision": "REQUIRE_HUMAN_REVIEW",
                    "risk_score": 0.55,
                    "risk_level": "MEDIUM",
                    "confidence": 0.65,
                    "analysis_result": {
                      "credit_analysis": "信用分数650，信用记录一般",
                      "income_analysis": "年收入8万，申请金额6万，需进一步评估",
                      "risk_factors": ["申请金额与收入比偏高", "信用分数中等"],
                      "approval_amount": 30000,
                      "recommended_rate": "年化利率7.5%"
                    },
                    "processing_info": {
                      "ai_model_version": "v1.2.0",
                      "processing_time_ms": 2800,
                      "workflow_id": "wf_20241201_002"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI决策结果处理成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "处理成功示例",
                    "value": {
                      "code": 0,
                      "message": "AI审批结果已成功处理",
                      "data": {
                        "application_id": "app_20241201_001",
                        "new_status": "AI_APPROVED",
                        "next_step": "等待放款审批"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AIAgentToken": []
          }
        ]
      }
    },
    "/api/v1/ai-agent/config/models": {
      "get": {
        "summary": "获取AI模型配置",
        "description": "获取当前可用的AI模型配置和规则参数",
        "operationId": "getAIModelConfig",
        "tags": ["AI智能体"],
        "responses": {
          "200": {
            "description": "成功获取模型配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "成功响应示例",
                    "value": {
                      "code": 0,
                      "message": "获取成功",
                      "data": {
                        "available_models": [
                          {
                            "model_id": "risk_model_v1",
                            "model_name": "风险评估模型V1",
                            "version": "1.2.0",
                            "status": "active"
                          }
                        ],
                        "risk_thresholds": {
                          "auto_approve_threshold": 0.3,
                          "auto_reject_threshold": 0.7
                        },
                        "decision_rules": {
                          "min_credit_score": 600,
                          "max_debt_ratio": 0.6,
                          "blacklist_auto_reject": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AIAgentToken": []
          }
        ]
      }
    },
    "/api/v1/ai-agent/workflow/trigger": {
      "post": {
        "summary": "触发AI审批工作流",
        "description": "触发AI审批工作流，开始自动化审批流程",
        "operationId": "triggerWorkflow",
        "tags": ["AI智能体"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["application_id"],
                "properties": {
                  "application_id": {
                    "type": "string",
                    "description": "申请ID"
                  },
                  "callback_url": {
                    "type": "string",
                    "description": "回调地址"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["high", "normal", "low"],
                    "default": "normal",
                    "description": "处理优先级"
                  }
                }
              },
              "examples": {
                "trigger_request": {
                  "summary": "触发工作流示例",
                  "value": {
                    "application_id": "app_20241201_001",
                    "callback_url": "http://localhost:8080/api/v1/applications/callback",
                    "priority": "normal"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "工作流触发成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "examples": {
                  "success": {
                    "summary": "触发成功示例",
                    "value": {
                      "code": 0,
                      "message": "AI审批工作流已成功触发",
                      "data": {
                        "workflow_id": "wf_20241201_001",
                        "status": "running",
                        "estimated_duration_ms": 5000
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AIAgentToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码，0表示成功"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "required": ["code", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误消息"
          },
          "error": {
            "type": "string",
            "description": "详细错误信息"
          }
        },
        "required": ["code", "message"]
      },
      "ApplicationInfo": {
        "type": "object",
        "properties": {
          "application_id": {
            "type": "string",
            "description": "申请ID"
          },
          "product_info": {
            "type": "object",
            "properties": {
              "product_id": {
                "type": "string",
                "description": "产品ID"
              },
              "name": {
                "type": "string",
                "description": "产品名称"
              },
              "max_amount": {
                "type": "number",
                "description": "最大额度"
              },
              "interest_rate_yearly": {
                "type": "string",
                "description": "年化利率"
              }
            }
          },
          "applicant_info": {
            "type": "object",
            "properties": {
              "user_id": {
                "type": "string",
                "description": "用户ID"
              },
              "real_name": {
                "type": "string",
                "description": "真实姓名"
              },
              "credit_score": {
                "type": "number",
                "description": "信用分数"
              },
              "annual_income": {
                "type": "number",
                "description": "年收入"
              }
            }
          },
          "amount": {
            "type": "number",
            "description": "申请金额"
          },
          "term_months": {
            "type": "integer",
            "description": "申请期限（月）"
          },
          "purpose": {
            "type": "string",
            "description": "申请用途"
          }
        }
      },
      "ExternalData": {
        "type": "object",
        "properties": {
          "credit_report": {
            "type": "object",
            "properties": {
              "score": {
                "type": "number",
                "description": "征信分数"
              },
              "level": {
                "type": "string",
                "description": "征信等级"
              },
              "detail": {
                "type": "string",
                "description": "征信详情"
              }
            }
          },
          "bank_flow": {
            "type": "object",
            "properties": {
              "average_monthly_income": {
                "type": "number",
                "description": "月均收入"
              },
              "debt_to_income_ratio": {
                "type": "number",
                "description": "债务收入比"
              },
              "account_stability": {
                "type": "string",
                "description": "账户稳定性"
              }
            }
          },
          "blacklist_check": {
            "type": "object",
            "properties": {
              "is_blacklisted": {
                "type": "boolean",
                "description": "是否在黑名单"
              },
              "risk_level": {
                "type": "string",
                "description": "风险等级"
              }
            }
          }
        }
      },
      "AIDecisionRequest": {
        "type": "object",
        "required": ["decision", "risk_score", "risk_level"],
        "properties": {
          "decision": {
            "type": "string",
            "enum": ["AUTO_APPROVED", "AUTO_REJECTED", "REQUIRE_HUMAN_REVIEW"],
            "description": "AI决策结果"
          },
          "risk_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "风险分数"
          },
          "risk_level": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH"],
            "description": "风险等级"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "置信度"
          },
          "analysis_result": {
            "type": "object",
            "properties": {
              "credit_analysis": {
                "type": "string",
                "description": "信用分析"
              },
              "income_analysis": {
                "type": "string",
                "description": "收入分析"
              },
              "risk_factors": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "风险因素"
              },
              "approval_amount": {
                "type": "number",
                "description": "建议审批额度"
              },
              "recommended_rate": {
                "type": "string",
                "description": "建议利率"
              }
            }
          },
          "processing_info": {
            "type": "object",
            "properties": {
              "ai_model_version": {
                "type": "string",
                "description": "AI模型版本"
              },
              "processing_time_ms": {
                "type": "number",
                "description": "处理时间(毫秒)"
              },
              "workflow_id": {
                "type": "string",
                "description": "工作流ID"
              }
            }
          }
        }
      },
      "AIModelConfig": {
        "type": "object",
        "properties": {
          "available_models": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "model_id": {
                  "type": "string",
                  "description": "模型ID"
                },
                "model_name": {
                  "type": "string",
                  "description": "模型名称"
                },
                "version": {
                  "type": "string",
                  "description": "模型版本"
                },
                "status": {
                  "type": "string",
                  "description": "模型状态"
                }
              }
            }
          },
          "risk_thresholds": {
            "type": "object",
            "properties": {
              "auto_approve_threshold": {
                "type": "number",
                "description": "自动审批阈值"
              },
              "auto_reject_threshold": {
                "type": "number",
                "description": "自动拒绝阈值"
              }
            }
          },
          "decision_rules": {
            "type": "object",
            "properties": {
              "min_credit_score": {
                "type": "number",
                "description": "最低信用分数"
              },
              "max_debt_ratio": {
                "type": "number",
                "description": "最大债务比例"
              },
              "blacklist_auto_reject": {
                "type": "boolean",
                "description": "黑名单自动拒绝"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "AIAgentToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AI Agent Token格式：'AI-Agent-Token your_token'"
      }
    }
  },
  "security": [
    {
      "AIAgentToken": []
    }
  ]
} 